0.2.0 / Main page version 1 complete
This version has most of the working parts complete, with a few exceptions that need a little bit of work here and there, including the newly planned side decorations that will be available in the next major version before release. Mostly to the headers and the burger menu. (that might not even exist.) There is also a few planned CTAs to make navigation and exploration of the page more dynamic and interactive.

The next few updates might contain semi-major page updates, and bugfixes. especialy for mobile.

+ added the loading screen
+ added the general main module that holds the header and footer of the page.
+ added the unformated parts of the main page to cards and card holder.
+ styled the page to make it easier to read, and to be all one unit.
~ made changes to the infini card holder to fix some bugs and make navigation more responsive and less prone to bugs.
~ made changes to the line module to support css
~ documented a few parts of the codebase